+++ tests/Neo.SmartContract.Framework.UnitTests/MapTest.cs
@@ -1,7 +1,6 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Neo.SmartContract.Testing;
 using Neo.SmartContract.Testing.TestingStandards;
-using Neo.VM;
 using Neo.VM.Types;
 using System.Reflection;
